pyproject.toml
[tool.poetry]
name = "solardb"
version = "0.0.1"
description = ""
authors = ["jadon <16850875+wolffshots@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
supabase = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
